requirements.txt
pytest~=7.2.2
pytest-bdd~=6.1.1
pyyaml~=5.4.1
invoke~=2.0.0